Revert "Implemented user authorization via jwt access token(path operation that refresh access token, not ready yet)"
This reverts commit ab51c3532e6ba04e24405bad63216caaa71cd3a0.
