+++ backend/app/api/auth/config.py
@@ -1,11 +1,9 @@
 """
-Config for auth service
+Config
 """
 from pydantic import BaseModel
 
 
 class AuthSettings(BaseModel):
-    jwt_algorithm: str
-    jwt_access_token_lifetime: int
-    jwt_refresh_token_lifetime: int
+    pass
