+++ backend/app/api/auth/schemas.py
@@ -4,11 +4,8 @@ Request bodies for path operations
 from pydantic import BaseModel, ConfigDict
 
 
-class UserDataResponse(BaseModel):
-    """
-    Model used for returning full user data
-    (can be used after creating/finding/updating user)
-    """
+class UserRegistrationResponse(BaseModel):
+    """Model used for returning created user"""
     model_config = ConfigDict(from_attributes=True)
 
     email: str
@@ -17,4 +14,3 @@ class UserDataResponse(BaseModel):
     age: int | None
     photo: str | None
     tel: str | None
-
