+++ backend/app/core/config.py
@@ -3,7 +3,6 @@ Settings for whole project
 """
 from pydantic_settings import BaseSettings, SettingsConfigDict
 from pydantic import BaseModel, computed_field
-from api.auth.config import AuthSettings
 
 
 class DatabaseSettings(BaseModel):
@@ -19,7 +18,11 @@ class Settings(BaseSettings):
 
     host: str = "localhost"
     port: int = 8000
-    secret_key: str | None = None
 
     db: DatabaseSettings
-    auth: AuthSettings
+    #auth: AuthSettings
+
+    @computed_field
+    @property
+    def uri(self) -> str:
+        return f"http://{self.host}:{self.port}"
